// frontend/src/types/generated/CalendarEventResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CalendarEventResponse = {
  type: string
  status: string
  repo_name: string
  hostname: string
  time: string
  report_id: number | null
  repo_id: number | null
  schedule_id: number | null
  archive_name: string | null
  error_message: string | null
}